backend/services/neo4j_client.py
"""
Neo4j client service.

Provides connection management and query execution for Neo4j.
"""

from django.conf import settings

# TODO: Implement Neo4j client
# from neo4j import GraphDatabase


class Neo4jClient:
    """Neo4j database client."""

    def __init__(self):
        self.uri = settings.NEO4J_URI
        self.user = settings.NEO4J_USER
        self.password = settings.NEO4J_PASSWORD
        self._driver = None

    # TODO: Implement connection methods
    # TODO: Implement query methods

