backend/requirements/dev.txt
-r base.txt

# Development
django-debug-toolbar>=4.2,<5.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0

# Code Quality
ruff>=0.1,<1.0
mypy>=1.7,<2.0
django-stubs>=4.2,<5.0

# Development Server
watchdog>=3.0,<4.0

